Change scrollbar callback #4 #9 #16

+++ src/main/java/me/do_you_like/mods/skinchanger/utils/gui/ModernGui.java
@@ -66,7 +66,7 @@ public abstract class ModernGui extends UILock implements UISkeleton {
 
     private List<InteractiveDrawable> selectedDrawables = Lists.newArrayList();
 
-    protected int yTranslation = 0;
+    protected float yTranslation = 0;
 
     @Override
     public final void initGui() {
